+++ src/test/java/lotto/LottoTest.java
@@ -1,21 +1,18 @@
-package lotto.domain;
+package lotto;
 
-import lotto.dto.Lotto;
-import lotto.dto.LottoProfit;
-import lotto.dto.LottoStatistics;
+import lotto.domain.Lotto;
 import org.junit.Test;
 
 import java.util.Arrays;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-
-public class LottoMachineTest {
+public class LottoTest {
 
     @Test
-    public void 당첨번호체크() {
+    public void 당첨숫자_카운트증가() {
         Lotto lotto = new Lotto(Arrays.asList(1, 2, 3, 4, 5, 6));
-        LottoMachine.checkWinningNumber(Arrays.asList(lotto), 1);
+        lotto.incrementMatchCount(3);
         assertThat(lotto.getMatchCount()).isEqualTo(1);
     }
 }
